# [mtouch] Add support for specifying how to optimize LLVM output.